OPR-107: Move failed database in own package
