+++ patient-reporter/src/main/java/com/hartwig/oncoact/patientreporter/PanelReporterApplication.java
@@ -3,7 +3,6 @@ package com.hartwig.oncoact.patientreporter;
 import java.io.File;
 import java.io.IOException;
 import java.time.LocalDate;
-import java.util.Map;
 
 import com.hartwig.lama.client.model.PatientReporterData;
 import com.hartwig.oncoact.patientreporter.lama.LamaJson;
@@ -13,8 +12,6 @@ import com.hartwig.oncoact.patientreporter.panel.PanelFailReport;
 import com.hartwig.oncoact.patientreporter.panel.PanelFailReporter;
 import com.hartwig.oncoact.patientreporter.panel.PanelReporter;
 import com.hartwig.oncoact.patientreporter.panel.QCFailPanelReportData;
-import com.hartwig.oncoact.patientreporter.qcfail.FailedDBFile;
-import com.hartwig.oncoact.patientreporter.qcfail.FailedDatabase;
 import com.hartwig.oncoact.patientreporter.reportingdb.ReportingDb;
 import com.hartwig.oncoact.util.Formats;
