+++ patient-reporter/src/main/java/com/hartwig/oncoact/patientreporter/panel/QCFailPanelReportData.java
@@ -2,12 +2,9 @@ package com.hartwig.oncoact.patientreporter.panel;
 
 import com.hartwig.oncoact.patientreporter.PanelData;
 
-import com.hartwig.oncoact.patientreporter.qcfail.FailedDatabase;
 import org.immutables.value.Value;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-
-import java.util.Map;
 
 @Value.Immutable
 @Value.Style(passAnnotations = {NotNull.class, Nullable.class})
